% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_functions.R
\name{all_caps}
\alias{all_caps}
\title{All Caps}
\usage{
all_caps(text)
}
\arguments{
\item{text}{string/ email subject line}
}
\description{
This function takes in any string and returns a boolean indicating
whether or not it is in all caps. This function is intended for use
on any of the email MEA datasets.
}
\examples{
emails_caps <- emails \%>\%
   mutate(all_caps = all_caps(text = subjectline))

}
