% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickbait.R
\docType{data}
\name{articles}
\alias{articles}
\title{Headlines from}
\format{
A data frame with 24666 rows and 3 variables:
\describe{
  \item{headline}{String}
  \item{clickbait}{Boolean}
  \item{ids}{Integer}
}
}
\source{
\url{XX}
}
\usage{
articles
}
\description{
XX
}
\keyword{datasets}
