% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles_clean.R
\docType{data}
\name{articles_clean}
\alias{articles_clean}
\title{Headlines cleaned}
\format{
A data frame with 22949 rows and 3 variables:
\describe{
  \item{headline}{String}
  \item{clickbait}{Boolean}
  \item{ids}{Integer}
}
}
\source{
\url{XX}
}
\usage{
articles_clean
}
\description{
This subset of the clickbait dataframe removes headlines containing
explicit language or suggestive content. Given the volume of headlines
containing such language (especially for clickbait == TRUE), this filtering
might not catch all problematic headlines. User discretion is advised.
}
\keyword{datasets}
