% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_functions.R
\name{begins_re}
\alias{begins_re}
\title{Re:}
\usage{
begins_re(text)
}
\arguments{
\item{text}{string/ email subject line}
}
\description{
This function takes in any string and returns a boolean indicating
whether or not it begins with "Re:", not followed by a space. This function is
intended for use on any of the email MEA datasets.
}
\examples{
\dontrun{
emails_punctuation <- emails \%>\%
   mutate(begins_re = begins_re(text = subjectline))
}

}
