% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{common}
\alias{common}
\title{Common Clickbait Phrases}
\format{
A data frame with 184 rows and 1 variable:
\describe{
  \item{phrases}{String}
}
}
\source{
\url{XX}
}
\usage{
common
}
\description{
XX
}
\keyword{datasets}
