% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_attributes.R
\name{common_phrases}
\alias{common_phrases}
\title{Common Phrases}
\usage{
common_phrases(ds)
}
\arguments{
\item{ds}{dataset}
}
\description{
This function takes in any dataframe with a headline variable and returns a
new dataframe containing the variable common_phrases, which is a boolean
that indicates the presence of one of the phrases found in the
dataframe common.
}
