% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_attributes.R
\name{contraction_words}
\alias{contraction_words}
\title{Contraction Words}
\usage{
contraction_words(ds)
}
\arguments{
\item{ds}{dataset}
}
\description{
This function takes in any dataframe with headline and ids variables
and returns a new dataframe containing the variable contraction_words,
which is an integer that enumerates the amount of contractions
(defined in dataset contractions) within a given headline.
}
