% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contractions.R
\docType{data}
\name{contractions}
\alias{contractions}
\title{Contractions}
\format{
A data frame with 71 rows and 1 variable:
\describe{
  \item{phrases}{String}
}
}
\source{
\url{XX}
}
\usage{
contractions
}
\description{
XX
}
\keyword{datasets}
