% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_load.R
\docType{data}
\name{emails}
\alias{emails}
\title{Email Subject Lines}
\format{
A data frame with 6,908 rows and 3 variables:
\describe{
\item{subjectline}{character Email subject line}
\item{type}{character Email classification into three levels: spam, hard_ham, and easy_ham}
\item{ids}{integer Row number}
}
}
\source{
\url{http://www.rdatasciencecases.org/Spam}
}
\usage{
emails
}
\description{
This dataset includes a set of email subject lines used for classification
of whether the message is spam (unsolicited commercial content) or not.
Many subject lines include  subject matter innapropriate for classroom use.
Given the volume of headlines containing such language
(especially for type == "spam"), user discretion is advised.
}
\keyword{datasets}
