% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_load.R
\docType{data}
\name{emails_test}
\alias{emails_test}
\title{Email Test}
\format{
A data frame with 1,382 rows and 3 variables:
\describe{
\item{subjectline}{character Email subject line}
\item{type}{character Email classification into three levels: spam, hard_ham, and easy_ham}
\item{ids}{integer Row number}
}
}
\source{
\url{http://www.rdatasciencecases.org/Spam}
}
\usage{
emails_test
}
\description{
This dataset is a random sample of 20\\% of the emails data, stratified by type.
}
\keyword{datasets}
