% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_load.R
\docType{data}
\name{emails_train}
\alias{emails_train}
\title{Email Train}
\format{
A data frame with 5,526 rows and 3 variables:
\describe{
\item{subjectline}{character Email subject line}
\item{type}{character Email classification into three levels: spam, hard_ham, and easy_ham}
\item{ids}{integer Original row number}
}
}
\source{
\url{http://www.rdatasciencecases.org/Spam}
}
\usage{
emails_train
}
\description{
This dataset is a random sample of 80\\% of the emails data, stratified by type.
}
\keyword{datasets}
