% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exaggerated.R
\docType{data}
\name{exaggerated}
\alias{exaggerated}
\title{Exaggerated Phrases}
\format{
A data frame with 346 rows and 1 variable:
\describe{
  \item{phrases}{String}
}
}
\source{
\url{XX}
}
\usage{
exaggerated
}
\description{
XX
}
\keyword{datasets}
