% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_attributes.R
\name{exaggerated_phrases}
\alias{exaggerated_phrases}
\title{Exaggerated Phrases}
\usage{
exaggerated_phrases(ds)
}
\arguments{
\item{ds}{dataset}
}
\description{
This function takes in any dataframe with a headline variable and returns a
new dataframe containing the variable exaggerated_phrases,
which is a boolean that indicates the presence of one of the phrases
found in the dataframe exaggerated.
}
