% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_functions.R
\name{has_ampersand}
\alias{has_ampersand}
\title{Ampersand}
\usage{
has_ampersand(text)
}
\arguments{
\item{text}{string/ email subject line}
}
\description{
This function takes in any string and returns a boolean indicating
whether or not there is an ampersand. This function is intended for use
on any of the email MEA datasets.
}
\examples{
\dontrun{
emails_dollar <- emails \%>\%
   mutate(has_ampersand = has_ampersand(text = subjectline))
}

}
