% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickbait_functions.R
\name{has_common_phrase}
\alias{has_common_phrase}
\title{Common Phrases}
\usage{
has_common_phrase(text)
}
\arguments{
\item{text}{string/ headline}
}
\description{
This function takes in any string and returns a boolean indicating
whether or not one of the phrases within the dataframe common
is present. This function is intended for use on any of the clickbait
MEA datasets.
}
\examples{
headlines_common <- headlines \%>\%
   mutate(has_common_phrase = has_common_phrase(text = title))

}
