% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_functions.R
\name{has_dear}
\alias{has_dear}
\title{Dear}
\usage{
has_dear(text)
}
\arguments{
\item{text}{string/ email subject line}
}
\description{
This function takes in any string and returns a boolean indicating
whether or not the word "dear" appears within it. This function is
intended for use on any of the email MEA datasets.
}
\examples{
emails_dear <- emails \%>\%
   mutate(has_dear = has_dear(text = subjectline))

}
