% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_functions.R
\name{has_dollar_sign}
\alias{has_dollar_sign}
\title{Dollar Sign}
\usage{
has_dollar_sign(text)
}
\arguments{
\item{text}{string/ email subject line}
}
\description{
This function takes in any string and returns a boolean indicating
whether or not there is a dollar sign. This function is intended for use
on any of the email MEA datasets.
}
\examples{
emails_dollar <- emails \%>\%
   mutate(has_dollar_sign = has_dollar_sign(text = subjectline))

}
