% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickbait_functions.R
\name{has_exaggerated_phrase}
\alias{has_exaggerated_phrase}
\title{Exaggerated Phrases}
\usage{
has_exaggerated_phrase(text)
}
\arguments{
\item{text}{string/ headline}
}
\description{
This function takes in any string and returns a boolean indicating
whether or not one of the phrases within the dataframe exaggerated is
present. This function is intended for use on any of the clickbait
MEA datasets.
}
\examples{
articles_exaggerated <- headlines \%>\%
   mutate(has_exaggerated_phrase =
          has_exaggerated_phrase(text = title))

}
