% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_functions.R
\name{has_mister}
\alias{has_mister}
\title{Mister}
\usage{
has_mister(text)
}
\arguments{
\item{text}{string/ email subject line}
}
\description{
This function takes in any string and returns a boolean indicating
whether or not the word "mister" or "mr" appears within it. This function
is intended for use on any of the email MEA datasets.
}
\examples{
emails_mister <- emails \%>\%
   mutate(has_mister = has_mister(text = subjectline))

}
