% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickbait_functions.R
\name{has_question_word}
\alias{has_question_word}
\title{Question Words}
\usage{
has_question_word(text)
}
\arguments{
\item{text}{string/ headline}
}
\description{
This function takes in any string and returns a boolean indicating
whether or not one of the phrases within the dataframe question is
present. This function is intended for use on any of the clickbait
MEA datasets.
}
\examples{
headlines_question <- headlines \%>\%
   mutate(has_question_word = has_question_word(text = title))

}
