% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_functions.R
\name{has_religious}
\alias{has_religious}
\title{Religious Subject Matter}
\usage{
has_religious(text)
}
\arguments{
\item{text}{string/ email subject line}
}
\description{
This function takes in any string and returns a boolean indicating
whether or not religious words are present. This function is
intended for use on any of the email MEA datasets.
}
\examples{
emails_religious <- emails \%>\%
   mutate(has_religious = has_religious(text = subjectline))

}
