% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines_load.R
\docType{data}
\name{headlines_test}
\alias{headlines_test}
\title{Headlines Test}
\format{
A data frame with 4589 rows and 3 variables:
\describe{
\item{title}{character Article headline}
\item{clickbait}{logical TRUE if the article is clickbait, and FALSE if not}
\item{ids}{integer Original row number}
}
}
\source{
\url{https://github.com/bhargaviparanjape/clickbait}
}
\usage{
headlines_test
}
\description{
This dataset is the remaining 20\\% of the observations
in "headlines" not found in the training set.
}
\keyword{datasets}
