% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines_load.R
\docType{data}
\name{headlines_train}
\alias{headlines_train}
\title{Headlines Train}
\format{
A data frame with 18360 rows and 3 variables:
\describe{
\item{title}{character Article headline}
\item{clickbait}{logical TRUE if the article is clickbait, and FALSE if not}
\item{ids}{integer Original row number}
}
}
\source{
\url{https://github.com/bhargaviparanjape/clickbait}
}
\usage{
headlines_train
}
\description{
This dataset is a random sample of approximately 80\\% of the observations
in "headlines", stratified by .
}
\keyword{datasets}
