% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_functions.R
\name{multiple_punctuation}
\alias{multiple_punctuation}
\title{Multiple Punctuation}
\usage{
multiple_punctuation(text)
}
\arguments{
\item{text}{string/ email subject line}
}
\description{
This function takes in any string and returns a boolean indicating
whether or not multiple punctuation marks are present. This function is
intended for use on any of the email MEA datasets.
}
\examples{
emails_punctuation <- emails \%>\%
   mutate(multiple_punctuation =
          multiple_punctuation(text = subjectline))

}
