% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickbait_functions.R
\name{num_contractions}
\alias{num_contractions}
\title{Contraction Words}
\usage{
num_contractions(text)
}
\arguments{
\item{text}{string/ headline}
}
\description{
This function takes in any string and returns the number of contractions
present.
This function is intended for use on any of the clickbait MEA datasets.
}
\examples{
headlines_contractions <- headlines \%>\%
   mutate(num_contractions = num_contractions(text = title))

}
