% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickbait_functions.R
\name{num_pronouns}
\alias{num_pronouns}
\title{Number of Pronouns}
\usage{
num_pronouns(text)
}
\arguments{
\item{text}{string/ headline}
}
\description{
This function takes in any string and returns an integer  equal to
the number of pronouns present.
This function is intended for use on any of the clickbait MEA datasets.
Pronouns detected include: I, me, you, he, him, she, her, it, we,
us, they, them, one, your, and my.
}
\examples{
articles_pronouns <- headlines \%>\%
   mutate(num_pronouns = num_pronouns(text = title))

}
