% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_attributes.R
\name{num_stop_words}
\alias{num_stop_words}
\title{Number of Stop Words}
\usage{
num_stop_words(ds)
}
\arguments{
\item{ds}{dataset}
}
\description{
This function takes in any dataframe with  headline and ids variables
and returns a new dataframe containing the variable num_stop_words,
which is an integer enumerating the number of stop words (as defined by
the tidytext package) within a given headline.
}
