% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickbait_functions.R
\name{num_words}
\alias{num_words}
\title{Number of Words}
\usage{
num_words(text)
}
\arguments{
\item{text}{string/ headline}
}
\description{
This function takes in any string and returns the number of words it has.
This function is intended for use on any of the clickbait MEA datasets.
}
\examples{
headlines_words <- headlines \%>\% mutate(num_words = num_words(title))

}
