% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_attributes.R
\name{positivity}
\alias{positivity}
\title{Positivity (AFINN) Score}
\usage{
positivity(ds)
}
\arguments{
\item{ds}{dataset}
}
\description{
This function takes in any dataframe with headline and ids variables and
returns a new dataframe containing the variable positivity,
which is an integer that represents the total AFINN score associated with
each headline. Scores of 0 represent neutral headlines, or headlines
not containing any words within the AFINN wordbank.
}
