% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\docType{data}
\name{question}
\alias{question}
\title{Question Phrases}
\format{
A data frame with 40 rows and 1 variable:
\describe{
  \item{phrases}{String}
}
}
\source{
\url{XX}
}
\usage{
question
}
\description{
XX
}
\keyword{datasets}
