% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_attributes.R
\name{question_words}
\alias{question_words}
\title{Question Words}
\usage{
question_words(ds)
}
\arguments{
\item{ds}{dataset}
}
\description{
This function takes in any dataframe with a headline variable and returns a
new dataframe containing the variable question_words, which is a boolean
that indicates the presence of one of the phrases found in the questions
dataset.
}
