% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_articles.R
\docType{data}
\name{sample_articles}
\alias{sample_articles}
\title{Sample Articles}
\format{
A data frame with 2000 rows and 3 variables:
\describe{
  \item{headline}{String}
  \item{clickbait}{Boolean}
  \item{ids}{Integer}
}
}
\source{
\url{XX}
}
\usage{
sample_articles
}
\description{
A small subset of of articles_clean. This dataset is intended to facilitate
more expedient modeling/ classification within the classroom. Clickbait and
non-clickbait headlines are equally represented within the randomly
generated sample.
}
\keyword{datasets}
