% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_attributes.R
\name{starts_with_num}
\alias{starts_with_num}
\title{Starts with Number}
\usage{
starts_with_num(ds)
}
\arguments{
\item{ds}{dataset}
}
\description{
This function takes in any dataframe with a headline variable and returns
a new dataframe containing the variable starts_with_num, which is a
boolean that indicates whether or not the headline starts with a number.
}
