% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_attributes.R
\name{words}
\alias{words}
\title{Number of Words}
\usage{
words(ds)
}
\arguments{
\item{ds}{dataset}
}
\description{
This function takes in any dataframe with headline and ids variables
and returns a new dataframe containing the variable words, which is an
integer that indicates the number of words within the headline.
}
