% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines.R
\docType{data}
\name{headlines_test}
\alias{headlines_test}
\title{Headlines Test}
\format{
A data frame with 4589 rows and 3 variables:
\describe{
  \item{title}{String}
  \item{clickbait}{Boolean}
  \item{ids}{Integer}
}
}
\source{
\url{https://github.com/bhargaviparanjape/clickbait}
}
\usage{
headlines_test
}
\description{
This dataset is a random sample of the remaining 20% of the observations
in "headlines" not found in the training set.
}
\keyword{datasets}
