% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickbait_functions.R
\name{num_stop_words}
\alias{num_stop_words}
\title{Number of Stop Words}
\usage{
num_stop_words(text)
}
\arguments{
\item{text}{string/ headline}
}
\description{
This function takes in any string and returns an integer equal to
the number of stop words present.
This function is intended for use on any of the clickbait MEA datasets.
}
\examples{
articles_stop_words <- headlines \%>\%
   mutate(num_stop_words = num_stop_words(text = title))

}
