% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headlines.R
\docType{data}
\name{sample_headlines}
\alias{sample_headlines}
\title{Sample Headlines}
\format{
A data frame with 2000 rows and 3 variables:
\describe{
  \item{title}{String}
  \item{clickbait}{Boolean}
  \item{ids}{Integer}
}
}
\source{
\url{https://github.com/bhargaviparanjape/clickbait}
}
\usage{
sample_headlines
}
\description{
This dataset is a random sample of 1000 clickbait article headlines and
1000 news article headlines from "headlines". This was created for use in
a guided activity using a data access platform called CODAP. If students
are interested in exploring how attributes differ across clickbait and news
headlines through use of 2 X 2 count tables, using this set will allow for
easier interpretation.
}
\keyword{datasets}
