% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickbait_functions.R
\name{starts_with_num}
\alias{starts_with_num}
\title{Starts with Number}
\usage{
starts_with_num(text)
}
\arguments{
\item{text}{string/ headline}
}
\description{
This function takes in any string and returns a boolean indicating
whether or not it begins with a number. This function is intended for
use on any of the clickbait MEA datasets.
}
\examples{
articles_starts_num <- headlines \%>\%
   mutate(starts_with_num = starts_with_num(text = title))

}
